Deserialization adjusts and insertion of InboundOrder

+++ src/main/java/com/mercadolibre/dambetan01/dtos/SectionDTO.java
@@ -18,5 +18,5 @@ public class SectionDTO {
     private String sectionCode;
 
     @NotNull(message = "warehouseCode is required")
-    private Integer warehouseCode;
+    private Long warehouseCode;
 }
